refactor: remove compatibility shims and canonicalize module ownership (P04)
- delete common/dataset_fingerprint.py (shim for contracts/dataset_fingerprint)
- delete core/release_reports.py (shim for contracts/release_reports)
- delete core/runtime_event.py (shim for contracts/runtime_event)
- delete policy/release_policy.py (shim for core/policy_engine)
- repoint promote.py, hosted_release_fixture_verifier.py, test_release_policy.py
  to import directly from core.policy_engine

Closes #122
