--- src/ml_lifecycle_platform/core/runtime_event.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-from ml_lifecycle_platform.contracts.runtime_event import RuntimeEvent
-
-__all__ = ["RuntimeEvent"]
