+++ src/ml_lifecycle_platform/policy/__init__.py
@@ -1,5 +1,1 @@
-from __future__ import annotations
-
-from .release_policy import PolicyDecision, Violation, evaluate_promotion_policy
-
-__all__ = ["PolicyDecision", "Violation", "evaluate_promotion_policy"]
+"""Release policy evaluation. Logic lives in core.policy_engine."""
